# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: PR Review API
  version: 0.1.0
  description: >
    Контракт Backend API для запуска автоматического ревью Pull Request
    и получения состояния и результатов.
servers:
  - url: /api/v1
    description: >
      Все пути ниже — относительно /api/v1. Префикс ставит само приложение, а не
      reverse-proxy: контракт не зависит от того, срезает ли прокси префикс.
      Фронт ходит на /api/v1/..., хостинг шлёт вебхуки на /api/v1/webhooks/{provider}.
paths:
...
  /webhooks/{provider}:
    parameters:
      - name: provider
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Provider"
    post:
      summary: Принять webhook-доставку от хостинга и поставить ревью в очередь
      operationId: receiveWebhook
      responses:
        "202":
          description: >
            Доставка принята: ReviewJob создан или, если для этого коммита уже есть
            нетерминальный, возвращён существующий. Дубль — не ошибка: не-2xx
            хостинг считает неудачной доставкой, повторяет её и после серии отказов
            может отключить вебхук, поэтому 409 здесь не отдаём.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewJob"
        "401":
          description: Подпись доставки не прошла проверку (ApiError.code = WEBHOOK_SIGNATURE_INVALID)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: >
            VCS недоступен или ответил ошибкой при получении диффа или метаданных;
            прогон не создан и задача не поставлена (ApiError.code = SCM_UNAVAILABLE).
            Ответ обязан спекой webhook-intake: сбой VCS — 502, а не прочий 5xx.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "501":
          description: >
            Провайдер есть в enum Provider, но приёмник вебхуков для него ещё
            не написан (ApiError.code = PROVIDER_NOT_SUPPORTED).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
...
  schemas:
    Provider:
      type: string
      enum:
        - github
        - gitlab
    TriggerSource:
      type: string
      enum:
        - webhook
        - manual
        - mention
...
    ReviewStatus:
      type: string
      enum:
        - queued
        - building_context
        - analysing
        - publishing
        - completed
        - failed
        - cancelled
    ReviewJob:
      type: object
      additionalProperties: false
      required:
        - id
        - provider
        - providerRepositoryId
        - pullRequestNumber
        - trigger
        - status
        - findingsCount
        - rejectedFindings
        - createdAt
        - updatedAt
        - lastProgressAt
        - error
      properties:
        id:
          type: string
          description: reviewRunId.
        provider:
          $ref: "#/components/schemas/Provider"
        providerRepositoryId:
          type: string
        pullRequestNumber:
          type: integer
          minimum: 1
        trigger:
          $ref: "#/components/schemas/TriggerSource"
        headCommitSha:
          type:
            - string
            - "null"
        baseCommitSha:
          type:
            - string
            - "null"
        status:
          $ref: "#/components/schemas/ReviewStatus"
        findingsCount:
          type: integer
          minimum: 0
        rejectedFindings:
          type: integer
          minimum: 0
        model:
          type:
            - string
            - "null"
        tokensUsed:
          type:
            - integer
            - "null"
          minimum: 0
        durationSeconds:
          type:
            - number
            - "null"
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastProgressAt:
          type: string
          format: date-time
          description: Обновляется на каждом переходе состояния; используется для выявления зависших прогонов (§7.5 PIPELINE_SPEC.md).
        error:
          oneOf:
            - $ref: "#/components/schemas/ReviewError"
            - type: "null"
...
    ReviewError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - retryable
        - scmProvider
        - providerAttempt
      properties:
        code:
          type: string
          enum:
            - SCM_AUTH_ERROR
            - PR_NOT_FOUND
            - REPOSITORY_NOT_FOUND
            - SCM_UNAVAILABLE
            - CONTEXT_PARSING_ERROR
            - LLM_RATE_LIMITED
            - LLM_TIMEOUT
            - LLM_UNAVAILABLE
            - LLM_INVALID_OUTPUT
            - REVIEW_RUN_STALE
            - INTERNAL_ERROR
        message:
          type: string
        retryable:
          type: boolean
        scmProvider:
          description: >
            Заполнено для SCM_AUTH_ERROR/SCM_UNAVAILABLE, иначе null. Называется
            scmProvider, а не provider: поле про сторону SCM, providerAttempt считает
            попытки и LLM, и SCM.
          oneOf:
            - $ref: "#/components/schemas/Provider"
            - type: "null"
        providerAttempt:
          type:
            - integer
            - "null"
          minimum: 1
          description: Попытка вызова LLM/SCM провайдера. Не путать с redelivery сообщения в очереди.
    ApiError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: >
            Закрытый список, как и ReviewError.code. Каждый ответ с ApiError называет
            свой код в description — это сверяет tests/test_docs.py.
          enum:
            - VALIDATION_ERROR
            - REVIEW_NOT_FOUND
            - REVIEW_ALREADY_ACTIVE
            - WEBHOOK_SIGNATURE_INVALID
            - SCM_UNAVAILABLE
            - PROVIDER_NOT_SUPPORTED
            - AUTH_CODE_INVALID
            - AUTH_REFRESH_INVALID
        message:
          type: string
...